readAllJsonFiles now can read the target directory from the argument.
